requirements.txt
psycopg2-binary==2.9.9
customtkinter==5.2.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0 